Create stack navigator for tab one
