+++ screens/WorkoutDetailsScreen.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { SafeAreaView, StyleSheet } from "react-native";
+
+export default function WorkoutDetailsScreen() {
+	return <SafeAreaView></SafeAreaView>;
+}
+
+const styles = StyleSheet.create({});
